EID-1666 Encrypt assertions in generated Response
The data within the assertions is already encrypted, however to remain
consistent with what we do elsewhere, such as sending the unsigned saml
response to the MSA, it makes sense to encrypt the new assertions too.

This has the benefit of allowing us to re-use the string transformers
which do the encryption for us, rather than pulling in a new one in
saml-lib.

To encrypt, we need to inject an `AssignableEntityToEncryptForLocator`
into the `RpAuthnResponseGeneratorService` and register the request ID
and entity to encrypt for.