EID-1714 Pass country response data to cycle3 attribute query
CountrySignedResponseContainer is now passed back in the various matching cycle states unless it's not present in which case an Optional.empty() is returned. This allows the original SAML Country Response to be sent with the Cycle3 data if required.

Co-authored by: Chris Wynne <christopher.wynne@digital.cabinet.office.gov.uk>