EID-1721 Pass countrySignedResponseContainer to MSA for user account creation
When the MSA doesn't find a user we enter the user account creation flow. In the case of unsigned assertions, we need to send the data from the `countrySignedResponseContainer` object as an assertion in the attribute query, rather than the original encrypted assertion.

Work has already been done to ensure that this object is available in the correct states, as well as the work to include the unsigned response assertion rather than the encrypted assertion (in HubEidasAttributeQueryRequestToSamlAttributeQueryTransformer).