BAU: Simplify use of generics
CertificateService did not need to declare generic types. Refactored to remove them.

This was driven by Jersey logging warnings that a concrete type for CertificateService could not be found. This seems to an issue with Jersey and didn't cause any real problems, but simplifying the generics here resolved the issue anyway.