BAU: Decrease the logging noise
We're logging a warn message when the S3 object is null. However, it will be null
most of the time since we only retrieve the object if it has changed.
The local cache expires every 30s by default, but the S3 file will not change that often.
This adds a condition to not log the message when it's happening during the reload phase
and we do not expect the full object to be retrieved.