EID-1762 Move eIDAS hash logging logic to SAML lib
We log hashes of eIDAS attributes in the hub as well as the proxy-node.
These are then compared in Splunk to ensure there's been no funny
business.

We'd like to consolidate all the logic for generating these hashes in
one place to ensure they're logged in the same way. This moves the logic
to saml-lib and pulls it in.