BAU Upgrade Gradle base image to 5.5.1-jdk
Gradle are decommissioning HTTP for Gradle services - see here for more
info: https://blog.gradle.org/decommissioning-http

A knock on effect of this is that if you're using an older version of
the JDK then there's a bug that'll kick a `peer not authenticated` error
when pulling dependencies.

This newer version of Gradle uses JDK 11.0.4. I didn't want to go beyond
this as we've seen issues in the hub with memory leaks when using 11.0.5
and are currently pinned at version 11.0.4, so this matches.