EID-1762: Upgrade to the latest version of saml-libs
Move logging of eIDAS attributes hash from `DecoratedSamlResponseToIdaResponseIssuedByIdpTransformer` to `IdpAuthnResponseTranslatorService`. This makes more sense as the service has all the information it needs to decide when it should log the hash and to call the logger. This also removes the clunky pattern of setting the logger conditionally in the transformer and avoids a potential bug where the transformer would log more often than necessary.

Author: @CharlesIC