BAU: Update to provide more details on exception
This change updates CertificateExpiryDateCheckService to catch exception inside the loop and to provide more details on the exception and certificate if it is available. This is to ensure that CertificateExpiryDateCheckService can continue to set certificate expiry dates metrics for the remaining certificates.

This change also updates logger to log exceptions as error messages rather than warning messages so that Sentry can catch them and alert 2nd Verify Line Support team.

Co-authored by: George Lund <george.lund@digital.cabinet-office.gov.uk>