HUB-413 Sentry event grouping is wrong
We have been using an old library to integrate DropWizard and Sentry.  Before we fine tune the way Dropwizard and Sentry work together, we're updating to the newer code that Sentry recommends.  Ideally this will solve all of our problems on its own and we won't have to do any further tweaking.  It probably won't but we definitely want to be using this before tweaking.

Change appender types from raven to sentry

Change format of `tags` configuration from `key:value` to `{"key": "value"}`.