HUB-545: Add an endpoint to config to provide a list of IDPs disconnected for registration
This is so the frontend can handle disconnecting IDPs correctly. It needs to know when an IDP has been disabled for registration so that it can show users a sensible error message if they manage to see that IDP shortly it gets disabled.

Also rename a few things related to the "registration" flow to make code easier to understand and make naming more consistent.