HUB-545: Include the `provide registration until` timestamp in the IPD DTO list from config
This is so the frontend can decide to hide disconnecting IDPs at an appropriate time to avoid a timing issue where the button to select an IDP get rendered shortly before the cut-off time and the user fails to press is before that cut-off time. In that case the user will see an error for a reason completely outside of their control.

Hiding IDPs disconnecting for registration a short period of time before the cut-off time is a crude solution to this as the Hub will still accept requests for those IDPs until the actual cut-off time. This doesn't eliminate the problem completely but given an appropriate time interval it reduces it greatly as most people are not likely to spend more than a few minutes on the picker.