Only send a [LOA2,LOA1] ordering of LOAs to IDPs that are enabled for registration.
We want IDPs which are enabled for receiving registration requests to be able to handle an Authn Request where the ordering of LOAs in the request is [LOA2, LOA1]. This is so that these IDPs will first attempt to achieve LOA2 for the User but failing that will try for LOA1. When a User is on the IDP in a sign-in journey they can switch to a registration journey, so it’s important that this ordering is used in both instances to ensure the IDP behaviour is consistent.

When a Service is configured with the [LOA2, LOA1] ordering of LOAs and they are sending a request to an IDP which is only configured for sign-in, then we want the request to contain just [LOA2]. This will be sent as a LOA2 request meaning a user who obtained LOA2 on registering can use the service. This will give the service the existing behaviour when sending requests to IDPs that are configured for only sign-in.

The change implemented here entails the following -

- Call config service to see if the IDP is enabled for registration.
- If it isn’t and the RP has the ordering of LOAs [LOA2, LOA1] then just use [LOA2].
- Add new tests to test this behaviour.