java: set sensible default for java dns cache
by default java caches DNS (seperatly from the OS DNS cache) queries
indefinitly. this is not a practical default when running in a cloud
environment with dynamic addresses of
load-balaners/applications/databases.

this changes the default java dns ttl to be 5s, leaving much of the
caching up to the OS as is expected.