BAU: Remove ruby
Github has flagged a security vulnerability with one of the dependencies
of the only gem we specify. This gem is only used by a ruby build script
for building packages. We no longer use the script (we use docker images
and the hub pipeline) so instead of fixing the vulnerability, lets just
remove ruby.

It also drops a couple of other unused scripts and updates the list of
dependencies in the README. I've removed node as a dependency as I can't
see where it's ever used. Happy to be shown wrong though.