HUB-814: Send VSP version info to Prometheus
We want to be able to easily see what version of the VSP services are
using.

This takes the version provided in authn requests and provides it to
Prometheus as a gauge.

The metric value of 1.0 is not the interesting part, as is only used as
a vehicle for the labels which describe the entity ID and the version.
It's only provided as metrics must provide a double value.