BAU Fix pseudo random failures
MatchingServiceRequestSenderTest fails sometimes.  We hypothesise that it's because the stubs used are @ClassRules and so static and so probably in the wrong state when
they get used sometimes for reasons not apparent to us.

Instead of diagnosing the problems properly and identifying where the problem lies, we've changed everything to non-static @Rules in the vain hope that this will fix all of our problems.

If we could reliably make the tests fail then it would be worth doing properly but we can't so we wouldn't know if it was actually fixed anyway.