BAU: Remove Ben Maines versions versions plugin
We don't seem to be using it anywhere other than the pre-commit.sh
script. We don't use pre-commit anywhere apart from locally.

The version we were using was from the Jenkins plugins repo, which has
now been removed (we no longer use Jenkins). Rather than doing the work
to add the version we want to our artifactory, we'd rather just remove
the plugin.