WIP
So far this has mostly involved changing the implementation of Base64 that's been used and updating some method signatures. The base64 stuff is in saml-libs I think which then means it's included as a test dependency.  I'm not in love with that but I'm treating this as a spike rather than a perfect solution so "meh? it works, let's do something better later."

There's a bit of Base64 that's been hardcoded in for the time being because it's contained in a separate library.

At the moment there are `method not found` type errors to do with DropWizard Modules - I suspect these are because something in the dependency tree is still requiring DropWiard 1.3 and the method signature we want is being hidden on the classpath.

If I didn't have other work to get on with, I'd search through the various ida-libs that we use - it's bound to be something like a test library to make it more annoying.