Upgrade dependencies to DW2
This sets out libraries to local versions - they've been upgraded to
DW2 but nit published yet. Will need updating if this all ever goes
live.

Opensaml has been bumped due to it pulling in an old version of an http
lib which was causing issues.

This drops dropwizard-guicier in favour of dropwizard-guicey. The former
hasn't been updated in over two years and doesn't seem to play well with
dropwizard two - not implementing certain interfaces. dropwizard-guicey
was inspired by guicier and seems to have more support and reasonable
docs.

We're also forcing httpclient to 4.5.13 as we were pulling in multiple
versions which were clashing on the classpath and causing issues.

jersey-media-jaxb is now needed - without it jersey can't seem to
serialize xml.

s3mock has been pushed to it's junit5 version to support the new style
of integration test.

We've started using junit5 so need to specify `useJUnitPlatform`