Let dropwizard-guicey register exception mappers.
They're picked up automatically as they're marked with the @Provider
annotation.

Apply the `@RequestScoped` annotation to all the exception mappers.
dropwizard-guicey gives them the singleton scope by default, which
prevents the @Context annotation working to inject request information

As of dropwizard 2, using @Context on a field or a setter doesn't work.
You need to provide it as a method parameter. This means adding the
Context parameters on the constructors of the sub-classes of the base
`PolicyExceptionMapper` and passing them through