Fix policy integration tests
Converts the policy integration tests to use the dropwizard-guicey Junit
5 extension, rather than class/app rules.

Also uses junit 5 to run the tests.

Update Policy exception mappers to use providers

There was an issue using @Context. The first time the mapper was
executed the request would be present. After that all it's attributes
would be null. @Context is confusing and is going away in future. Using
a provider is cleaner and means we don't need to add the @RequestScope
annotation.

This also converts the RedisTestRule, which was used by a class of
tests, from a Junit 4 rule to a Junit 5 extension, so that it can be
used.