Make saml-engine integration tests works
Adds the appropriate annotations to exception mappers.

Converts all integration test to use dropwizard-guicey junit5 extension.

Changes the ConfigStubRule to ConfigStubExtension so it works with
junit5.

I had to ensure all entity ID's used in
RpAuthnResponseGeneratorResourceTest were unique to avoid saml-engines
internal cache for config, which was preventing the stub from being hit.
There might be a better way to manage this.